Chat GPT: Introduce a base class
